Wrap Flask-Foundation assets inside Flask-Assests
In order to allow all assets to be served from the same location,
Flask-Foundation's assets are now being handled through Flask-Assets. In
debug mode, Flask-Foundation's non-minified CSS and JavaScript will be
used. When `ASSETS_DEBUG` is false, Flask-Assets will minimize and
concatenate Flask-Foundation's assets.

Along with this change, the home page template and base layout are being
brought up to date with the splash page. This includes adding a setting
to control the Google Analytics profile.