Add initial page describing the CFP
People will have questions about the CFP process. It's important to
provide answers to as many of those equestios as we can. The text of
this page still needs to be proofread, and I'm sure it will change a few
times before it's finalized, but this is a good start.