Prompt users to enter a name and bio
Currently users can register for the site and immediately submit a talk
proposal. For the accepted talks we want to show speakers' names on the
schedule. By forcing users to the settings page, they will have to fill
in their profile before submitting a proposal.

For new users, they will now be directed immediately to the settings
page. Perhaps they should be redirected to more appropriate pages after
filling in the fields.