Add support for event date ranges
The `Event` class has several date ranges associated with it. It has the
`is_registration_active` property to check whether or not registration
for the event is open. It's now getting the
`is_call_for_proposals_active` property to check whether or not the CFP
is active and accepting submissions.

Both of these properties are now being utilized to help populate the top
nav. `Event.is_call_for_proposals_active` is also being utilized to
prevent users from submitting new proposals to events outside of the
specified date range. Existing proposals can still be edited after the
close of the CFP.