Add some additional sponsor information
In addition to providing a tiny bit of information to potential
sponsors, this change also adds a call to action to get them over to the
application page.

The sponsorship levels should really be driven by the `Level` model.
That will need a few additional fields added to it.

Also the email address may change to a specific person's.

References #16