Limit supported HTML tags for reStructuredText
Because users are allowed to enter text that will be converted to HTML,
we need to be wary of the tags that are allowed. We don't want anyone to
override styles on the page or inject (harmful) JavaScript. The `bleach`
library can be used to help limit what it's allowed.

`bleach` can be a little too aggressive though, so the list of supported
tags needs to be expanded to included headings and paragraphs.