Use local time
When entering times through the admin, it's natural for people to do so
using their local time (in the case of everyone working on PyGotham
that's Eastern). While UTC could be entered, there are places where the
date is displayed on the site. To allow everyone to always think in
local time, Arrow is being introduced. Fortunately SQLAlchemy-Utils has
an `ArrowType` that can be used in place of `DateTimeFIeld`s.