Use Level to drive sponsorship prospectus
Rather than hardcoding all of the details about the sponsorship levels
in the sponsorship prospectus template, the `Level` model can be used to
drive the page. To accomplish this the cost needs to be added to the
model.

Additionally, by adding the desired limit on the number of sponsors at a
level, we can automatically indicate when a level is sold out. Using the
`cached-property` library means the sponsor count doesn't need to be
recalculated each time the `is_sold_out` property is referenced in the
template.