Add Sponsors page
This is a page that will display the sponsors for the event. The
sponsors will be grouped by level and displayed alphabetically.

In addition to adding the page, a new property, `accepted_sponsors`, is
being added to the `Level` model. It servers as a shortcut to the
sponsors confirmed for that level, something needed in multiple places.

Closes #16