Create anchors to jump to specific sponsors
Providing direct links to sponsors gives us the chance to use it in
announcements (e.g., on Twitter) as well as gives the sponsors a chance
to show off their sponsorship.

This is accomplished by generating a slug from each sponsor's name and
placing an anchor inside the `<li>`. Currently the slug is generated on
demand. In the future this may be handled as a database field.

Closes #36