Add sponsors to the home page
Sponsors are now being displayed on the home page. They are positioned
to the left of the announcements, which is now an aside. In the case of
small viewports, the announcements appear below the sponsors.

To accomplish this, a `get_accepted` function is being added to the
`sponsors` package. It behaves similarly to the `news.get_active`
function.