Add talk list and detail pages
Two new views are being added, `talks.index` and `talks.detail` to
display the talk list and detail view of a talk, respectively. Accepted
talks will be listed alphabetically.

The accepted talks can be accessed through a new property on `Event`
called `accepted_talks`.

In order to allow talks to be marked as accepted without them appearing
on the site ahead of time, a new column is being added to `Event` called
`talk_list_begins`. It will contain the date and time when the talk list
will first appear on the site. There is also an associated property
being added named `talks_are_published`. It is being used to control the
presence of the talk list in the menu, the accepted label in the user
dashboard, and whether or not the new views are accessible.