Add bulk accept and reject for talks
In order to add custom actions to an admin view, `ActionxMixin` needs to
be a base class for the view class. The `model_view` function doesn't
allow for custom inheritance, and getting it to work would have been
tricky because the views require a method to go with each action.

To get this behavior to work, `TalkModelView` is being changed to a
class (as the name always suggested) with `talk_model_view` becoming the
instantiated version. It provides views to either accept or reject a
batch of talks.