Add speaker profiles
In addition to adding the view and template for the speaker profiles,
the `User` model is getting a couple of properties. The first property,
`accepted_talks`, returns all of a user's accepted talks. The second
property, `has_accepted_talks`, utilizes the first to indicate whether
or not a user has accepted talks.