Show sponsorship applications on dashboard
After a sponsor has submitted an application, they may want to update
some details. This new form will let them update the contact information
by first visiting their dashboard.

To help make the dashboard a little easier on the eye, some of the
styles being used on the page are being updated. The most important
thing to know about these changes is that there will be more of a
distinction between the styles of buttons and labels.