Add the talk schedule page
These are the workings of a schedule. It includes the models, template,
and admin.

The general flow is that an event has days. Each day is made up of
slots. Slots are linked to rooms and presentations. Finally,
presentations are linked to talks.

A slot can override its content with text instead of linking to a
presentation.

While no special styles have been defined yet, the kind of a slot can be
used to override the styles for alike slots.