Handle empty slots
It is possible that there will be blank slots in the schedule. When this
happens, failing to set the slot's `content_override` to one or more
spaces will cause the page to break, raising `UndefinedError`. By
checking for the presence of a presentation associated with the slot
first, the exception can be avoided and the page will render correctly.