Add presentation times to talk detail page
The times of a talk's presentation should be available when viewing its
detail page. This is especially important because slot's end times are
not available on the schedule page.

In addition to adding the times to the page, the string representation
of `Slot` are being adjusted slightly to both include the end time and
read a little easier.

Some style work needs to be done to make the time look nicer on the
page.

Closes #62