Don't allow presentations without talks
As part of making the schedule page handle empty slots better, the
`Presentation.talk_id` field is no longer allowed to be null.
`talk_id` was allowed to be null because `content_override` originally
belonged to `Presentation`. Now that it is part of `Slot`, a
presentation has no reason not to have a talk.