Add favicon
A favicon is being added to the site. Because it is a PNG, Chrome,
Safari, Firefox, Opera, and IE 11+ will display it.

In addition to the favicon, a "touch" icon is being added for iOS and
Android, as well as a Metro tile for IE 10+.

The favicon is 96x96, the touch icon is 152x152, and the Metro tile is
144x144.