Add the recording release
While the talk submission for referenced the recording release, we never
actually had one before. This new endpoint and template expose one,
based on the one used by the PSF for PyCon.

A link to the release is being added to the talk submission form.

Note: The release says that the checkbox excepting the release will
be checked by default. This isn't actually the case. It should be fixed
at some point -- an issue will be created -- but since the CFP is over,
it isn't necessary now.

Closes #77