Convert About pages to RST content in database
Removes all about hard-coded About content in favor of storing it in
the database via the admin site. As a side effect, introduces database
calls to (acceptably) load content on appropriate pages and
(unfortunately) render the About menu in the navbar. Addresses several
issues in Github: fixes #84, makes obsolete #24, #59, #83, and #86.

---

Add method to autogenerate and display navbar links

Links are generated by inspecting the url_map of the running
application. Additional links may be added (or existing ones replaced)
by implementing a get_navbar_links function in a module. Before being
passed to the template, the links are normalized, converted to title
case, and sorted alphabetically.