Add management command to create admin users
One of the biggest points of confusion when setting up a new
installation of this project is how to create an admin user. The process
requires using the `create_user` management command and then issuing SQL
statements to create the admin role and add it to the new user.

This new management command takes care of all of that.

Because of the overlap in functionality between the `create_user` and
`create_admin` commands, a new class, `RegisterUserMixin`, is being
added.