Make the recording release opt-out
The wording used on the page says the release is opt-out. This makes the
functionality match.

The change has to be implemented in this way because SQLAlchemy doesn't
use a column's default values until it's inserted into the database. The
`recording_release` column will start out with an initial value of
`None`. This means that even though WTForms-Alchemy will use a column's
default value, the `None` will override any default and result in the
checkbox not being checked.

Closes #97