Change how new talks are added to the session
Flask-SQLAlchemy turned on autoflush by default. With the upgrade in
9bb3c58, the new talk submission page started raising an exception when
trying to flush the session. There really is no need to add the talk to
the session until it's been populated. Unfortunately associating it with
an `Event` meant that it needed a session.

By passing `event_id` to `Talk()` instead of `event`, no session is
necessary until we're ready to insert the talk.