Move talk durations to their own model
While it was nice using an enum for talk durations, the needs of the
field have changed. The durations available for 2015 are different than
those in 2014. While the enum's values could be changed, this would
result in one of two things:

1. 2014 values would be removed, breaking the 2014 event.
2. 2015 would include values not wanted, leading to lots of confusion
   for users.

Neither of these are good. By moving durations to a separate model, old
values can be hidden (through the `inactive` field) and old events would
still function.

To accomplish the switch, a migration of both the schema and data is
being included. The migration will add the new table and a foreign key
from the talks table, populate the durations table based on the enum's
values, populate the foreign key with correct values, remove support for
nulls from the foreign key field, and remove the enum. All of these
actions can be rolled back.

Closes #132