Add a slug to events
In order to support multiple simultaneous events or the ability to view
previous events, there needs to be a way to identify an event, most
likely through a URL. A slug field is being added to the `Event` model
to add this ability. A migration is being provided that will add the
field and populate it using a slugified version of the event's name.

References #118