Allow AboutPage slugs to be set manually
If the slug changes whenever an `AboutPage`'s title changes, URLs
bookmarked by users or indexed by search engines can break. There may
also be times when a slug that doesn't exactly match the title may be
desired.

By only setting the slug when it's empty and exposing the field in the
admin, this level of control is achieved.