Apply new font and set header styles
The font being used is Titillium Web. It is available from Google Fonts.

In addition to the font, styles are being set for h1 and h2 tags in the
main content. A media query is being used to only use rounded corners
when rendering on a desktop. The login, logout, and register buttons are
being changed to work the same way, through a media query rather than a
class.