Apply more style changes
In addition to changes to the CSS, there one other change being made.
The sponsors page is being updated because nested `.row` elements have a
weird behavior. The `div.row` for each level is being changed to a
`section.level`. This is a bit more semantic and allows for more direct
targeting.