Properly handle reStructuredText headers
The default functionality when converting reST to HTML treats document
and section headers with the same height. This leads to them all being
h1 tags. Our desired behavior is for them to be h2 tags. The behavior is
being changed by passing in a setting override.

Along with the new section header behavior comes div tags with a
`section` class. div tags aren't allowed in our use of bleach, so they
render as output. This makes it look like something is broken. Instead
of converting these tags to rendered output, they're now going to be
stripped.