Include AboutPage's title in the filtered content
In order to properly treat heading in the page's content as section
headers, a document header is needed. This comes in the form of the
page's title. Rather than making people include the title in the page's
content when it's entered in the admin, a new property is being added to
`AboutPage` to provide the full document's contents. This value can then
be passed to the `rst` filter for processing.

The h1 tag is being removed from the template as it will now be rendered
through the filtered document.