Don't include other event's sponsors
There are a couple of places where we look up sponsors for the current
event. The current query uses an outer join to join the sponsors and
levels tables. This happens when no explicit join is added to the query.
An explicit join is being added (`join(Level)`) to switch to an inner
join.

This will affect both the home page and the sponsors page.