Include the event slug in Flask-Security URLs
In order for the Flask-Security endpoints to load without problems
loading the current event, they need to be exposed through URLs that
contain the event's slug. Unfortunately this value is normally provided
to a blueprint when it's defined. Fortunately Flask-Security has a
setting, `SECURITY_URL_PREFIX`, to include a prefix to all URLs.