Make Flask-Security redirect to the correct root
With the introduction of the current event's slug to URLs, `/` no longer
serves as the root of the site. While the production environment gets
around this by redirecting `/` to `/2015/`, this doesn't apply to
development environments. It makes for a bad experience for people who
are trying to contribute. Fortunately Flask-Security has settings that
allow you to specify the names of endpoints for it to redirect to after
logging in and out.