Remove fields from talk proposal form
A few fields are being removed from the talk proposal form: type, target
audience, and objectives.

We're only accepting talk submissions this year. We probably want to add
some way of saying which types are available for each event, but until
then, it needs to go.

This work was done by @djds23 in 21540c0, but was reverted to put all of
the other fields back.