Improve descriptions of fields for talk proposal
Some of the fields in the talk proposal form were confusing. The hope is
that adding better descriptions to the fields will help. To further
this, the additional requirements field is being rebranded as the
additional notes field. Hopefully this will better express that this
field is meant to given additional information to the program committee
rather than the attendees.