Add create_event management command
This adds a new management command called `create_event`. It handles
setting the fields that are required for an event to be functional. This
is especially important now that the current event is specified through
the URL. Without this command, there would be no way to access the
site's admin without creating the first event in the database directly.

Closes #150