Add Twitter handle to user profiles
This is a new, optional field that will be part of the user profile. It
uses a custom field and widget to handle leading `@` symbols entered by
the user (or an admin). The handle will be stored in the database
without the symbol but will be displayed in forms with it.

Closes #148