Add sponsors to the home page
This adds an event's sponsors to the home page. The logic is being added
to the template and some styles are being added. A couple of styles are
also being added for the sponsors page. Normally I would do this as two
separate commits, but they are closely related enough that I'm making an
exception.