Handle the home page post-CFP
With the current version of the home page template, the CFP open date is
shown whenever the CFP isn't active. That means it shows up after the
CFP closes. This will fall back on a message about registering.

The real fix here is to add announcements to the home page and feature
the most recent, but I want to make sure something will be in place
Monday morning after the CFP is over.