Add calls to action
The original version of the 2015 home page featured a call to action
block. Unfortunately there was no way to control the call to action
outside of changing the template. A new model is being added to allow
for them to be controlled through the admin.

There will be some overlap between calls to action and announcements,
but calls to action will likely be kept around longer than an
announcement will be the most recent.