Fix positioning of the footer
When a page's content is too short, we push the footer to the bottom of
the viewport. Unfortunately, because the top nav isn't including in the
main content, setting the content's height to the size of the viewport
results in the footer being pushed down the height of the top nav. In
order to fix this, the top nav needs to be moved inside the div with the
100% height. `.top-nav` and `#content` are being wrapped in a new div,
and the height and margin styles are being moved from `#content` to this
new div.

On a personal note, I dislike using these divs as, semantically, the top
nav and footer should be at the same level of the DOM. That's why I gave
the new div an annoying identifier that explains its purpose a little.