Persist proposals information on dashboard
Rather than removing the proposals section from the dashboard of users
who haven't submitted a talk after the CFP has ended, a note will be
displayed to encourage them to submit next year.

Conversely, should we ever launch a new event before the CFP opens, a
message will be displayed to inform the user that the CFP hasn't yet
opened.