Initial skeleton for the API
This needs work, but it lays out a decent starting point to expose
information through an API. Much of the schema needs fleshing out, more
endpoints should be exposed, and nested fields should be used more. That
said, it should be suitable enough as a starting point for future work
or just a one-time dump of data for pyvideo.