Hide presentation when schedule isn't published
No information about the schedule should be publicly available until the
schedule is published. Because of the rush to get the schedule live on
the site for PyGotham 2014, some corners were cut.

A talk's presentation slot information should only be available on the
detail page once the schedule for the event has been published.