Use only start times for generating times list
Using both start times and end times causes issues with rowspans when
generating the displayed talk schedule. By using only the start time to
generate the list of times (equated to rows in the display) and using
the closest time greater than or equal to the talk's end time to
generate that talk entry's rowspan, the cell is created with the correct
size.