Pad schedules times for pairwise iteration
When iteration over a schedule's times, we do so two items (start and
end) at a time. Since the list was being built with start times only,
the final time slot for each day was being omitted as the last start
time was used solely as an end time.

The final end time is being added to the end of the list so that the
correct number of pairwise iterations will happen.