Allow images in About pages
In order to display images on the "static" About pages, we need to tell
bleach that `img` tags are supported. We don't want this to always be
the case when converting from reStructuredText to HTML, though. To
handle this, a new parameter is being added to the `rst` filter that
accepts a list of tags to allow. The parameter is then being utilized in
the template for the About pages to allow `img`.

`img` tags will support the `src`, `alt` and size-related attributes.

Closes #233