Adjust the alignment of the sponsor headings
Having all of a sponsor's information left-aligned causes the logo and
name to look slightly misaligned. This look better centered. For 2014,
all of the sponsor information was centered. The description looks
better left-aligned, though, so paragraph tags are left to the left.
This led to removing the paragraph tag around the URL, which was only
done for spacing purposes anyway.