Activate events added through management command
When marking an event as active, it won't be fully activated until the
activity begins date. The management command to add new events wasn't
set up to do this, which could be confusing for some people. The command
will now also set the `activity_begins` field to the current time when
its being marked as active.