Update Nginx for PyGotham 2016
PyGotham 2016 marks the first time we're going to use year-specific
subdomains. While this marks the third such approach -- 2014 was
pygotham.org and 2015 was pygotham.org/2015/ -- it should be the last.
It will give us the most flexibility moving forward, especially as we
move to the new Docker and CoreOS-based infrastructure.

The most notable difference in behavior as far as Nginx is concerned is
that, rather than redirecting the root path to the current event, the
zone apex (a.k.a. the root domain) will be redirected to the current
event. Because this redirect will change for each new PyGotham, the
redirect is being handled as a 302.