Move the time zone to a setting
When we had `'America/New_York'` written in one place, it wasn't so bad.
As can be seen from this patch, the time zone is used in a bunch of
different places. Moving it to a setting means we don't have to worry
about typing it wrong.

UTC is being used as the default because it's the least insane time
zone, even though America/New_York will be the most likely time zone to
use.

Closes #277