Support top-level navbar about pages
Currently, we use `AboutPage.title` for two purposes, one of which being
navbar placement. This change removes that usage from the code and
replaces `navbar_section` with a more flexible `navbar_path`.

A note on indexing the query used to fetch `AboutPage`s and building the
navbar:  adding an index on `array_length(navbar_path, 1)` could
theoretically speed this query up, but 1) there are plenty of other ways
to make this faster (it should probably not run before every single
request, it can be cached), and 2) my gut feeling is that because we
have such a small number of rows in this table, a sequential scan of the
table will almost always be run, so it provides no value at this time
(this may be reevaluated in the future).