Add Code of Conduct & Privacy Policy to new events
Once upon a time, all of the "static" pages were stored inside the
templates folder of the `frontend` application. This had the advantage
of being able to see the change history of documents such as the Code of
Conduct and Privacy Policy. This disadvantage was that applications with
multiple events would have the same copy for the pages.

0418bda introduced the `AboutPage` model. This model was used to move
many of the "static" pages out of the templates and into the database.
This not only allowed events to have their own version of page, but
events could have their own set of pages. It also introduced the
advantage of being able to update the copy on a page without deploying
code.

The disadvantage here is that it becomes harder for people to track
changes we make to the aforementioned documents. Given their nature, the
more transparent we can be able them, the better.

This changeset not only gives us the ability to begin tracking the
changes we make to them, but it also allows us to prepopulate them when
adding new events through the `create_event` management command.