Fix schedule rendering
SQLAlchemy 1.0 introduced a change that wraps query results containing
fields from multiple tables in a "result" collection. This causes
problems for us when we try to assign a rowspan to each slot.
Downgrading to a version of SQLAlchemy prior to 1.0 fixes this.

We should look into a solution that will allow us to upgrade again, but
since we want to publish the schedule soon, this will have to do for
now.