Add subdomain support to the API
Because of the `SERVER_NAME`, the API currently does not work. As a
short term hot fix, this change gets the API back into a working state.
In the long term, we should either use the event identified by the
subdomain in all API requests or split the API application out into a
separately deployed application (e.g. `https://api.pygotham.org/`).