Use Marshmallow for serialization
Our use of Flask-Restful's marshalling functionality is hackish. There
exists significant support for moving Flask-Restful's internal
marshalling to Marshmallow [1], so we should switch to that.
Flask-Marshmallow is used for convenience functions (e.g. `jsonify`).

This changeset is strictly a port from Flask-Restful's builtin
marshalling mechanism to Marshmallow. TODOs and FIXMEs still exist and
should be addressed in a separate change.

Flask-Marshmallow relies on changes to `flask.jsonify` made in Flask
0.11, so our version must be upgraded in order to use it. 0.11.1 is the
latest stable release of Flask.

[1] https://github.com/flask-restful/flask-restful/issues/335