Pin (sort of) the version of flake8-docstrings
`pep257` was renamed to `pydocstyle`. Part of the initial release under
the new name is a change that treats classes nested inside other classes
as public as long as they aren't named with a leading underscore.
`flake8-docstrings` 1.0 is the first release to use the new name.

This causes linting errors with this codebase because `WTForms-Alchemy`
and `Flask-RESTful` both used a nested class named `Meta` to describe
their outer class. Until we figure out how we want to handle this going
forward, the version of `flake8-docstrings` is being restricted to
anything prior to 1.0 so that other changes can be made and merged after
their builds pass.