Add format hints to the create_event command
Whenever I use the `create_event` command (typically this only happens
when I set up a new development instance of PyGotham) I forget that the
CFP start date field requires a datetime. This causes the command fails
to create the event, requiring me to run the command again and enter all
of the information all over again. By providing format hints as part of
the prompt, people won't need to know the underlying data model to know
how to enter the values for dates and datetimes.

Signed-off-by: Andy Dirnberger <andy@dirnberger.me>