Install correct major version of jQuery
Foundation's bower-specific release contains a version requirement for
jQuery that specifies being greater than or equal to 2.1. jQuery 3.0
contains a change that is incompatible with Foundation. There is an
upstream commit that fixes the version requirement that pins jQuery to
2.1.x.

The sha of the commit is being specified in the bower file to force it
to install a version of Foundation that will install the correct version
of jQuery.

(This could also be fixed by upgrading the version of Foundation, now
called `foundation-sites`, but that's potentially a much larger change.)

Signed-off-by: Andy Dirnberger <andy@dirnberger.me>