2017 colors and logos
These are the new colors and logos for PyGotham 2017. The event name and
dates are included in the logo, so the text is being removed from the
template and an `alt` tag is being added to the `img`.

Due to being at PyCon, there isn't an SVG of the text-less version of
the logo available so a PNG is being used for the time being.

Signed-off-by: Andy Dirnberger <andy@dirnberger.me>